+++ src/app/js/app.js
@@ -2,7 +2,7 @@
   'use strict';
 
   angular.module('orrportal', [
-    'ngRoute',
+    'ui.router',
     'ngSanitize',
     'ui.bootstrap'
     ,'ui.select'
@@ -27,7 +27,13 @@
     .constant("cfg", appConfig)
     .run(init)
     .run(xeditable)
-    .config(routes)
+    .config(uiRoutes)
+    .run(['$rootScope', '$state', '$stateParams', function ($rootScope, $state, $stateParams) {
+      if (appUtil.debug) {
+        $rootScope.$state = $state;
+        $rootScope.$stateParams = $stateParams;
+      }
+    }])
   ;
 
   init.$inject = ['$rootScope', '$location', 'rUri', 'cfg'];
@@ -85,82 +91,98 @@
     editableOptions.theme = 'bs3';
   }
 
-  routes.$inject = ['$routeProvider', 'rUri'];
+  uiRoutes.$inject = ['$stateProvider', '$urlRouterProvider', 'rUri'];
+  function uiRoutes($stateProvider, $urlRouterProvider, rUri) {
 
-  function routes($routeProvider, rUri) {
-    console.log("routes: rUri=", rUri);
-    $routeProvider
-      // root "/" route set depending on whether we have a URI to dispatch:
-      .when('/', rUri
-        ? {
+    console.debug("uiRoutes: rUri=", rUri);
+
+    $urlRouterProvider.otherwise("/");
+
+    $stateProvider
+      .state('/', rUri
+          ? {
+          url: '/',
           //templateUrl: 'js/uri/views/uri.tpl.html',
           //controller: 'UriController'
           template: '<orr-ont></orr-ont>'
-          }
-        : {
+          //templateUrl: 'js/ont/views/ont.tpl.html',
+          //controller: 'OntController'
+        }
+          : {
+          url: '/',
           templateUrl: 'js/main/views/main.tpl.html',
           controller: 'MainController'}
       )
 
-      .when('/newvoc/', {
-          template: '<orr-ont></orr-ont>'})
-
-      .when('/so/:so*', {
-        templateUrl: 'js/main/views/main.tpl.html',
-        controller: 'MainController'})
-
-      .when('/st/', {
-        templateUrl: 'js/st/views/st.tpl.html',
-        controller: 'SearchTermsController'})
-      .when('/st/:st*', {
-        templateUrl: 'js/st/views/st.tpl.html',
-        controller: 'SearchTermsController'})
-
-      .when('/kw/', {
-        templateUrl: 'js/kw/views/kw.tpl.html',
-        controller: 'KeywordSearchController'})
-      .when('/kw/:kw*', {
-        templateUrl: 'js/kw/views/kw.tpl.html',
-        controller: 'KeywordSearchController'})
-
-      .when('/uri/:uri*', {
-        templateUrl: 'js/uri/views/uri.tpl.html',
-        controller: 'UriController'
+      .state('newvoc', {
+        url: '/newvoc',
+        template: '<orr-ont></orr-ont>'
+        //templateUrl: 'js/ont/views/ont.tpl.html',
+        //controller: 'OntController'
       })
 
-      .when('/org/:orgName*', {
-        templateUrl: 'js/org/views/org.tpl.html',
-        controller: 'OrgController'
-      })
-
-      .when('/neworg', {
-        templateUrl: 'js/org/views/createOrg.tpl.html',
-        controller: 'CreateOrgController'
-      })
-      .when('/user/:userName*', {
-        templateUrl: 'js/user/views/user.tpl.html',
-        controller: 'UserController'
-      })
-
-      .when('/signIn', {
-        templateUrl: 'js/auth/views/login.tpl.html',
-        controller: 'LoginController'
-      })
-      .when('/signIn/:redirect*', {
-        templateUrl: 'js/auth/views/login.tpl.html',
-        controller: 'LoginController'
-      })
-
-      .when('/fireauth-test', {  // TODO remove
-        templateUrl: 'js/fireauth/views/test.tpl.html'
-      })
-
-      .when('/rx', {
+      .state('rx', {
+        url: '/rx',
         templateUrl: 'js/upload/views/sequence.tpl.html',
         controller: 'UploadController'
       })
 
-      .otherwise({redirectTo: '/'});
+      .state('neworg', {
+        url: '/neworg',
+        templateUrl: 'js/org/views/createOrg.tpl.html',
+        controller: 'CreateOrgController'
+      })
+
+      .state('searchOnt', {
+        url: '/so/{so:.*}',
+        templateUrl: 'js/main/views/main.tpl.html',
+        controller: 'MainController'})
+
+      .state('searchTerm', {
+        url :'/st/{st:.*}',
+        templateUrl: 'js/st/views/st.tpl.html',
+        controller: 'SearchTermsController'})
+
+      .state('searchKw', {
+        url: '/kw/{kw:.*}',
+        templateUrl: 'js/kw/views/kw.tpl.html',
+        controller: 'KeywordSearchController'
+      })
+
+      .state('uri', {
+        url: '/uri/{uri:.*}',
+        templateUrl: 'js/uri/views/uri.tpl.html',
+        controller: 'UriController'
+      })
+
+      .state('org', {
+        url: '/org/{orgName:.*}',
+        templateUrl: 'js/org/views/org.tpl.html',
+        controller: 'OrgController'
+      })
+
+      .state('user', {
+        url: '/user/{userName:.*}',
+        templateUrl: 'js/user/views/user.tpl.html',
+        controller: 'UserController'
+      })
+
+      .state('signIn', {
+        url: '/signIn',
+        templateUrl: 'js/auth/views/login.tpl.html',
+        controller: 'LoginController'
+      })
+      .state('signInRedirect', {
+        url: '/signIn/{redirect:.*}',
+        templateUrl: 'js/auth/views/login.tpl.html',
+        controller: 'LoginController'
+      })
+
+      .state('fireauth-test', {  // TODO remove
+        url: '/fireauth-test',
+        templateUrl: 'js/fireauth/views/test.tpl.html'
+      })
+    ;
   }
 
-})();
+  })();